pyproject.toml
[project]
name = "system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "eventlet>=0.40.0",
    "flask>=3.1.1",
    "flask-socketio>=5.5.1",
    "line-bot-sdk>=3.17.1",
    "openai>=1.90.0",
    "openai-whisper>=20240930",
    "pyaudio>=0.2.14",
    "pyautogui>=0.9.54",
    "pydub>=0.25.1",
    "pypdf2>=3.0.1",
    "python-pptx>=1.0.2",
    "realtimestt>=0.3.104",
    "rich>=14.0.0",
    "simple-websocket>=1.1.0",
    "torch>=2.6.0",
    "torchaudio>=2.6.0",
]

# sources from https://blog.csdn.net/Humbunklung/article/details/146046406
[tool.uv.sources]
torch = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true
